+++ assets/css/aspire-update.css
@@ -13,21 +13,7 @@
 }
 
 .aspirepress-settings-field-hosts-wrapper .aspirepress-settings-field-hosts-row {
-    display: flex;
-    justify-content: flex-start;
-    flex-wrap: wrap;
-    gap: 20px;
-}
-
-.aspirepress-settings-field-hosts-wrapper .aspirepress-settings-field-hosts-row .aspirepress-settings-field-hosts-left,
-.aspirepress-settings-field-hosts-wrapper .aspirepress-settings-field-hosts-row .aspirepress-settings-field-hosts-right {
-    flex: 0 1 auto;
     margin: 0 0 10px;
-    box-sizing: border-box;
-}
-
-.normal-mode .advanced-setting {
-    display: none !important;
 }
 
 #aspirepress-generate-api-key {
