+++ assets/js/aspire-update.js
@@ -58,21 +58,13 @@ class FieldRules {
                 if (jQuery(this).is(':checked')) {
                     FieldRules.show_field('api_key');
                     FieldRules.make_required('api_key');
-                    FieldRules.show_field('api_host');
-                    FieldRules.show_field('rewrite_wporg_api');
-                    FieldRules.show_field('api_url');
-                    FieldRules.show_field('rewrite_wporg_dl');
-                    FieldRules.show_field('api_download_url');
+                    FieldRules.show_field('api_hosts');
                     FieldRules.check_enabled_rewrite_wporg_api.init();
                     FieldRules.check_enabled_rewrite_wporg_downlods_api.init();
                 } else {
                     FieldRules.hide_field('api_key');
                     FieldRules.remove_required('api_key');
-                    FieldRules.hide_field('api_host');
-                    FieldRules.hide_field('rewrite_wporg_api');
-                    FieldRules.hide_field('api_url');
-                    FieldRules.hide_field('rewrite_wporg_dl');
-                    FieldRules.hide_field('api_download_url');
+                    FieldRules.hide_field('api_hosts');
                 }
             }).change();
         }
@@ -111,16 +103,10 @@ class FieldRules {
             jQuery('#aspirepress-settings-field-enable_debug').change(function () {
                 if (jQuery(this).is(':checked')) {
                     FieldRules.show_field('enable_debug_type');
-                    FieldRules.show_field('exclude_debug_type');
-                    FieldRules.show_field('debug_log_path');
                     FieldRules.show_field('disable_ssl_verification');
-                    FieldRules.show_field('examine_responses');
                 } else {
                     FieldRules.hide_field('enable_debug_type');
-                    FieldRules.hide_field('exclude_debug_type');
-                    FieldRules.hide_field('debug_log_path');
                     FieldRules.hide_field('disable_ssl_verification');
-                    FieldRules.hide_field('examine_responses');
                 }
             }).change();
         }
