--- classes/AspirePress_ApiWordpressOrgRewriteRule.php
@@ -1,11 +1,0 @@
-<?php
-
-class AspirePress_ApiWordpressOrgRewriteRule implements AspirePress_RewriteRuleInterface
-{
-    use AspirePress_RewriteRuleTrait;
-
-    public function __construct(string $apiDestination)
-    {
-        $this->setHostRewriteRule('api.wordpress.org', $apiDestination);
-    }
-}
