--- classes/AspirePress_DownloadsWordpressOrgRewriteRule.php
@@ -1,11 +1,0 @@
-<?php
-
-class AspirePress_DownloadsWordpressOrgRewriteRule implements AspirePress_RewriteRuleInterface
-{
-    use AspirePress_RewriteRuleTrait;
-
-    public function __construct(string $apiDestination)
-    {
-        $this->setHostRewriteRule('downloads.wordpress.org', $apiDestination);
-    }
-}
