--- classes/AspirePress_HostRewriterInterface.php
@@ -1,6 +1,0 @@
-<?php
-
-interface AspirePress_HostRewriterInterface
-{
-    public function rewrite($url) : string;
-}
