+++ includes/class-aspirecloud-api.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * The Class for Interacting with Aspire Cloud API.
+ *
+ * @package aspire-update
+ */
+
+namespace Aspirepress;
+
+/**
+ * The Class for Interacting with Aspire Cloud API.
+ */
+class AspireCloud_API {
+	/**
+	 * TODO.
+	 */
+}
