--- sample_api_response.json
@@ -1,1 +1,0 @@
-{"api_key":"1012c79c930814ba6f7fa0b699fb667320e007a42893a1a1"}
